Rework where clause (ConditionClause) syntax
The current syntax in 0.0.3 is a bit rough with the mostly-mandatory
`:and/:or`, especially when moving from one argument to multiple. I find
myself constantly forgetting `:and`, then hitting a runtime error, then
having to change it. This also made it harder to align the syntax
between `where` and `join...on`.

To resolve this, the `:and/:or` flag is gone, and `:and` is always
assumed. This means that going from one condition to two conditions does
not have any other change that you must remember.

I found `:and` logic to be much more common in my code so it's the
default. To specify `:or` logic, use a subgroup.